add dry_run flag

+++ delete-all-tweets.py
@@ -23,17 +23,23 @@ def twitterconnect():
 
 def main():
     args = parse_args()
+    dry_run = args.dry_run
     api = twitterconnect()
 
     for page in tweepy.Cursor(api.user_timeline).pages():
         for tweet in page:
-            print(f"Deleting : {tweet.text}")
-            api.destroy_status(tweet.id)
+            if dry_run:
+                print(f"Not deleting : {tweet.text}")
+            else:
+                print(f"Deleting : {tweet.text}")
+                api.destroy_status(tweet.id)
 
 
 def parse_args():
     parser = argparse.ArgumentParser(description='delete-all-tweets')
     parser.add_argument('--debug', help="Display debugging information", action="store_const", dest="loglevel", const=logging.DEBUG, default=logging.INFO)
+    parser.add_argument('-d', '--dry_run', help="Dry run (doesn't delete anything)", dest='dry_run', action='store_true')
+    parser.set_defaults(dry_run=False)
     args = parser.parse_args()
     logging.basicConfig(level=args.loglevel)
     return args
